Add color-preference widget in player screen
This _will_ be incomplete, because the metadata scraper only operates
when there is high confidence - if multiple cards match the search, or
if a slang name is used, or if there are multiple commanders (or a
background), no colour will be set. Still - will be cool to get it in
front of folks and see what they think.